requirements.txt
viam-sdk
hidapi
co2meter
